JS: Make Date.getTime return Double instead of Int. #KT-4141 Fixed

+++ js/js.libraries/src/core/date.kt
@@ -1,5 +1,5 @@
 package kotlin.js
 
 public external class Date() {
-    public fun getTime(): Int
+    public fun getTime(): Double
 }
